+++ database/migrations/2021_07_27_171043_edit_comments.php
@@ -14,7 +14,7 @@ class EditComments extends Migration
     public function up()
     {
         Schema::table('comments', function (Blueprint $table) {
-            $table->integer('rating');
+            $table->unsignedTinyInteger('rating');
         });
     }
